Move Deadline to utils.py
